Adicionado endpoints e correções para a integração com front

+++ src/main/java/br/com/tiinforma/backend/services/aws/StorageService.java
@@ -17,6 +17,7 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -24,7 +25,9 @@ import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.time.LocalDate;
+import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @Slf4j
@@ -43,32 +46,65 @@ public class StorageService  {
     private UsuarioRepository usuarioRepository;
 
 
-    public Video uploadFile(
+    @Transactional
+    public String uploadFile(
             MultipartFile file,
             String titulo,
             String descricao,
             String categoria,
-            LocalDate dataPublicacao,
-            List<String> palavraChave,
+            LocalDate dataCadastro,
+            String palavraChaveString,
             Criador criador
     ) {
+        if (file.isEmpty()) {
+            throw new IllegalArgumentException("O arquivo não pode estar vazio");
+        }
+
+        String contentType = file.getContentType();
+        if (contentType == null || !contentType.startsWith("video/")) {
+            throw new IllegalArgumentException("Tipo de arquivo inválido. Por favor, envie um vídeo.");
+        }
+
         File fileObj = convertMultiPartFileToFile(file);
         String fileName = System.currentTimeMillis() + "_" + file.getOriginalFilename();
 
-        s3Client.putObject(new PutObjectRequest(bucketName, fileName, fileObj));
-        fileObj.delete();
+        try {
+            ObjectMetadata metadata = new ObjectMetadata();
+            metadata.setContentType("video/mp4");
 
-        Video video = Video.builder()
-                .titulo(titulo)
-                .descricao(descricao)
-                .categoria(categoria)
-                .palavraChave(palavraChave)
-                .dataPublicacao(dataPublicacao != null ? dataPublicacao : LocalDate.now())
-                .key(fileName)
-                .criador(criador)
-                .build();
+            PutObjectRequest putObjectRequest = new PutObjectRequest(bucketName, fileName, fileObj);
+            putObjectRequest.setMetadata(metadata);
 
-        return videoRepository.save(video);
+            s3Client.putObject(putObjectRequest);
+
+
+            List<String> palavrasChaveList = null;
+            if (palavraChaveString != null && !palavraChaveString.trim().isEmpty()) {
+                palavrasChaveList = Arrays.stream(palavraChaveString.split(","))
+                        .map(String::trim)
+                        .collect(Collectors.toList());
+            }
+
+            log.info("Salvando vídeo com palavras-chave: {}", palavrasChaveList);
+            Video video = Video.builder()
+                    .titulo(titulo)
+                    .descricao(descricao)
+                    .categoria(categoria)
+                    .palavraChave(palavrasChaveList != null ? String.join(",", palavrasChaveList) : null)
+                    .dataPublicacao(dataCadastro != null ? dataCadastro : LocalDate.now())
+                    .key(fileName)
+                    .criador(criador)
+                    .build();
+            log.info("Video a ser salvo: {}", video);
+
+            videoRepository.save(video);
+
+            return "File uploaded and video saved with key: " + fileName;
+        } finally {
+            if (fileObj.exists()) {
+                fileObj.delete();
+            }
+        }
     }
